backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.3
python-dotenv==1.0.0
web3==6.5.0
sqlalchemy==2.0.19
eth-account==0.8.0
eth-utils==2.1.0
eth-abi==4.0.0
requests==2.31.0
numpy==1.23.5
torch==2.0.1
tensorflow-cpu==2.12.0
pytest==7.3.1
pytest-asyncio==0.21.0
ipfshttpclient==0.7.0
black==23.7.0
isort==5.12.0
mypy==1.4.1
asyncio==3.4.3
